.Net: Serialize python code execution result (#11969)
### Motivation, Context and Description

This PR changes the format of Python code execution results from a
string, which is the result of concatenating property values, to the
product of serializing the result property values. The idea is that AI
models can reason well about JSONs but may not reason as effectively
over custom strings.