.Net: Update to Intermediate KernelFunction Approach (#12055)
### Motivation and Context

This change introduces a `IntermediateKernelFunction` that allows a
`KernelFunction` Name property to override the original `AIFunction`
name behavior. This simplifies the usage of `AIFunction`s as
`KernelFunctions` where the concept of `PluginName` exists and don't
need to be mixed with the `AIFunction.Name`.

The above approach also allows `KernelFunction`s to be used as
`AIFunction`s where the `PluginName` is automatically reflected as a
prefix in the `AIFunction.Name`.

## Example
```
kernelFunction.Name -> "MyFunction"
kernelFunction.PluginName -> "MyPlugin"

((AIFunction)kernelFunction).Name -> "MyPlugin_MyFunction"
```