.Net: Rename DeleteCollection -> EnsureCollectionDeleted (#12113)
I'm mainly proposing this rename to align with colleciton creation,
which has been renamed to EnsureCollectionCreatedAsync (next to which
we'd have EnsureCollectionDeletedAsync). It's slightly longer than
DeleteCollectionAsync, but the consistency here seem to be worth it.

I'm also noting that EF also has an EnsureDeleted/EnsureCreated pair
([docs](https://learn.microsoft.com/en-us/ef/core/managing-schemas/ensure-created)).

Note that this changes both the method on VectorStoreCollection and on
VectorStore (where there's no EnsureCreated).