.Net: Contextual function provider (#12254)
### Motivation, Context, and Description

This PR adds functionality that allows AI agents to advertise functions
relevant to the current context, rather than advertising all available
functions.

The functionality is implemented by two components:
- The `FunctionStore` class is responsible for the vectorization of
functions, upserting them into the provided vector store, and performing
vector searches. This class is internal at the moment as it's used only
by the `ContextualFunctionProvider` class. Later, it might be promoted
to public if required.
- The `ContextualFunctionProvider` class delegates function
vectorization and persistence to the `FunctionStore`. It vectorizes the
current context and delegates the vector search to the `FunctionStore`
class again.

Logging will be added in follow-up PRs.

Contributes to:
https://github.com/microsoft/semantic-kernel/issues/10074

---------

Co-authored-by: westey <164392973+westey-m@users.noreply.github.com>