.Net: Migrate sln to slnx (#12173)
I'm planning some project/filesystem organization for the MEVD projects
to make everything easier to manage (and more standard). Before doing
that, I thought I might as well submit a modernization from the old SLN
format to the new, simplified SLNX format:
https://devblogs.microsoft.com/visualstudio/new-simpler-solution-file-format.
This should be a completely transparent change (all tooling/VS/VS Code
support this), but please give it a try and let me know if you see any
trouble.

Note that we had an odd "Publish" configuration (in addition to the
traditional Debug/Release), some projects were configured for it in the
SLN, others weren't; this is probably just a bit of mess that
accumulated over time. As part of this PR I cleaned that up to leave a
tight, minimal and standard SLNX. This should make changes easier to
track and conflicts much easier to manage.

Let me know what you think!

---------

Co-authored-by: Dmytro Struk <13853051+dmytrostruk@users.noreply.github.com>
Co-authored-by: Roger Barreto <19890735+RogerBarreto@users.noreply.github.com>