Python: Add special notes to samples that are still using `AgentGroupchat` (#12377)
### Motivation and Context

<!-- Thank you for your contribution to the semantic-kernel repo!
Please help reviewers and future users, providing the following
information:
  1. Why is this change required?
  2. What problem does it solve?
  3. What scenario does it contribute to?
  4. If it fixes an open issue, please link to the issue here.
-->
As we move away from `AgentGroupChat` to the new orchestrations, we want
to make sure it's clear to developers getting started with SK are aware
of the changes.

### Description

<!-- Describe your changes, the overall approach, the underlying design.
These notes will help understanding how your code works. Thanks! -->
We are including a special note to all the samples that use
`AgentGroupChat` to inform developers that this feature is no longer
maintained and they should consider using `GroupChatOrchestration`
instead.

We are keeping the samples for now so that developers can still refer to
them and we don't break any external links to these samples, but we want
to make sure they are aware of the changes and can migrate to the new
orchestrations.


### Contribution Checklist

<!-- Before submitting this PR, please make sure: -->

- [x] The code builds clean without any errors or warnings
- [x] The PR follows the [SK Contribution
Guidelines](https://github.com/microsoft/semantic-kernel/blob/main/CONTRIBUTING.md)
and the [pre-submission formatting
script](https://github.com/microsoft/semantic-kernel/blob/main/CONTRIBUTING.md#development-scripts)
raises no violations
- [x] All unit tests pass, and I have added new tests where possible
- [x] I didn't break anyone :smile: