.Net: Allow hyphens in function name (#12389)
### Motivation, Context and Description

This PR allows hyphens in function names for seamless integration with
MCP tools and M.E.AI functions that support hyphens in their names.

Out of the scope of the PR is support for functions with hyphens in SK
prompt templates, which will need to be extended to support hyphens
when/if needed.