.Net: Fix Labels Implementation in Vertex AI to Use Dictionary<string, string> (#12636)
### Description
This pull request addresses issue #12633 by updating the Labels property
in the `GeminiPromptExecutionSettings` and `GeminiRequest` classes to
use a `Dictionary<string, string>` instead of a string. This change
aligns the implementation with the Vertex AI API requirements, which
expect labels to be sent as key-value pairs in JSON format.

Fixes #12633

---------

Co-authored-by: Adit Sheth <adsheth@microsoft.com>
Co-authored-by: Roger Barreto <19890735+rogerbarreto@users.noreply.github.com>