.Net: Prepare code for A2A SDK and AsyncEnumerable NuGet update (#12842)
### Motivation, Context and Description
This PR prepares the SK codebase for a future update of the A2A agent to
the new A2A .NET SDK and establishes the foundation for easy adoption of
the System.Linq.AsyncEnumerable package when it released. This will
completely eliminate the need for the AsyncEnumerable polyfills that SK
currently uses.

As part of this preparation, this PR stops using of the
System.Linq.Async package and replaces all SK calls to its APIs with
either equivalent calls to the AsyncEnumerable polyfills or in-place
implementations. This prevents "The call is ambiguous between..." name
conflicts that would occur when the AsyncEnumerable APIs from the A2A
.NET SDK's transitive dependency (System.Linq.AsyncEnumerable) become
available to the Agents.UnitTests project, which already contains
equivalent APIs from either AsyncEnumerable polyfills and/or the
System.Linq.Async package brought in from other projects that have the
Agents.UnitTests project specified as "InternalVisibleTo".

As a result, almost all non-production code (tests and samples) uses
AsyncEnumerable APIs from the preview version of the
System.Linq.AsyncEnumerable package, while production code uses either
AsyncEnumerable APIs from polyfills or uses in-place implementations.