.Net: Allow the MongoDB VectorData tests to run without an image (#12860)
### Motivation and Context

We are running the SK tests in our
https://github.com/mongodb-labs/ai-ml-pipeline-testing repo and would
like to also be able to run them against live MongoDB Atlas clusters for
further verification rather than just the Docker image.

### Description

The PR adjusts the vector data tests (only as per @roji recommendation)
so that if an environment variable named `MONGODB__CONNECTIONURL` exists
then that is used to connect to MongoDB rather than using TestContainers
to fire up a Docker image of MongoDB Atlas.

Additionally to improve reliability it was found that we need to
increase the number of attempts to wait for the data to be ready after
creating an index. (Bumped from 20 to 200). We're going to look at
adding index status helper methods in the C# Driver going forward for a
better experience.

Once this is done we should be able to fully enable the new tests on our
repo and diagnose any further instabilities and then be able to submit a
PR to fully-enable the image-based tests here.

### Contribution Checklist

<!-- Before submitting this PR, please make sure: -->

- [x] The code builds clean without any errors or warnings
- [x] The PR follows the [SK Contribution
Guidelines](https://github.com/microsoft/semantic-kernel/blob/main/CONTRIBUTING.md)
and the [pre-submission formatting
script](https://github.com/microsoft/semantic-kernel/blob/main/CONTRIBUTING.md#development-scripts)
raises no violations
- [x] All unit tests pass, and I have added new tests where possible
- [x] I didn't break anyone :smile: