.Net: [MEVD] More test cleanup (#13320)
Another batch of test cleanup work - this almost concludes what I have
planned for the tests.

* We no longer create any collections with GUID names; all collections
have a manually-defined name, and are deleted when a test starts
running. This ensures that no collections are left behind if a test
crashes or is stopped in the middle, and makes it easier to examine the
database.
* Removed global SimpleRecord and SimpleFixture, each test suite now has
its own record and fixture.
* Cleaned up and renamed DistanceFunctionTests to only focus on testing
the different distance functions, and added corresponding tests
elsewhere.
* Added IndexKindTests to specifically cover different index types.