Fix C# 14 Contains translation for MEVD providers (#13407)
This fixes the failing Contains-related MEVD filter tests. I'm not quite
sure what happened since I merged #13263 (at which point these were
passing) - possibly #13395 - but these pass again.

Note that I still plan to go back and design the translators in a way
which doesn't require the code to be duplicated, but we shouldn't block
on that.