.Net: Switch MEVD MongoDB Driver to v3.51 (#13370)
### Motivation and Context

Switches the MongoDB Driver to the latest 3.5 release which had a number
of breaking changes.

This affects both the MongoDB vector data project and the CosmosMongoDB
project.

This fixes #11652 and likely addresses  #12707 and partly #10291.

### Description

Switches the MongoDB driver to 3.5. Part of the breaking changes is that
GUIDs in BSON no longer have a default storage format specified due to
the need to switch from the C#-only format to the cross-MongoDB-driver
standard format. Setting this is achieved in this PR by way of both a
convention for the registry based mode and an alternative to
BsonValue.Create in the scenarios where we don't have access to
conventions/serialization such as key creation and using the dynamic
mapper.

### Contribution Checklist

- [x] The code builds clean without any errors or warnings
- [x] The PR follows the [SK Contribution
Guidelines](https://github.com/microsoft/semantic-kernel/blob/main/CONTRIBUTING.md)
and the [pre-submission formatting
script](https://github.com/microsoft/semantic-kernel/blob/main/CONTRIBUTING.md#development-scripts)
raises no violations
- [x] All unit tests pass, and I have added new tests where possible
- [x] I didn't break anyone :smile:


cc @roji

---------

Co-authored-by: Arthur Vickers <ajcvickers@hotmail.com>
Co-authored-by: Mark Wallace <127216156+markwallace-microsoft@users.noreply.github.com>
Co-authored-by: westey <164392973+westey-m@users.noreply.github.com>
Co-authored-by: Roger Barreto <19890735+rogerbarreto@users.noreply.github.com>
Co-authored-by: Shay Rojansky <roji@roji.org>